fix: biginteger

+++ app/db/models.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, Boolean
+from sqlalchemy import Column, Boolean, BigInteger
 from app.db.base import Base
 
 
 class ChatEntry(Base):
     __tablename__ = "chats"
 
-    chat_id = Column(Integer, primary_key=True)
+    chat_id = Column(BigInteger, primary_key=True)
     show_intro = Column(Boolean, nullable=False)
     only_active = Column(Boolean, nullable=False)
